+++ src/main/java/com/equilibrium/item/tools_attribute/metal/MetalShovel.java
@@ -147,7 +147,7 @@ public class MetalShovel extends ToolItem implements AdditionalAttribute{
         return AdditionalAttribute.super.xpCost(toolMaterial, durabilityLevel);
     }
     @Override
-    public int maxPlayerDurabilityBoost(ToolMaterial toolMaterial,PlayerEntity player){
+    public double maxPlayerDurabilityBoost(ToolMaterial toolMaterial, PlayerEntity player){
         return AdditionalAttribute.super.maxPlayerDurabilityBoost(toolMaterial,player);
     }
