+++ src/main/java/com/equilibrium/item/tools_attribute/metal/MetalSword.java
@@ -89,7 +89,7 @@ public class MetalSword extends ToolItem implements AdditionalAttribute {
         return AdditionalAttribute.super.xpCost(toolMaterial, durabilityLevel);
     }
     @Override
-    public int maxPlayerDurabilityBoost(ToolMaterial toolMaterial,PlayerEntity player){
+    public double maxPlayerDurabilityBoost(ToolMaterial toolMaterial, PlayerEntity player){
         return AdditionalAttribute.super.maxPlayerDurabilityBoost(toolMaterial,player);
     }
